// src/up-api/models/ListTransactionsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Up API
 * The Up API gives you programmatic access to your balances and transaction data. You can request past transactions or set up webhooks to receive real-time events when new transactions hit your account. It’s new, it’s exciting and it’s just the beginning. 
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ListAccountsResponseLinks } from './ListAccountsResponseLinks';
import {
    ListAccountsResponseLinksFromJSON,
    ListAccountsResponseLinksFromJSONTyped,
    ListAccountsResponseLinksToJSON,
} from './ListAccountsResponseLinks';
import type { TransactionResource } from './TransactionResource';
import {
    TransactionResourceFromJSON,
    TransactionResourceFromJSONTyped,
    TransactionResourceToJSON,
} from './TransactionResource';

/**
 * Successful response to get all transactions. This returns a paginated
 * list of transactions, which can be scrolled by following the `prev` and
 * `next` links if present.
 * @export
 * @interface ListTransactionsResponse
 */
export interface ListTransactionsResponse {
    /**
     * The list of transactions returned in this response.
     * @type {Array<TransactionResource>}
     * @memberof ListTransactionsResponse
     */
    data: Array<TransactionResource>;
    /**
     * 
     * @type {ListAccountsResponseLinks}
     * @memberof ListTransactionsResponse
     */
    links: ListAccountsResponseLinks;
}

/**
 * Check if a given object implements the ListTransactionsResponse interface.
 */
export function instanceOfListTransactionsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "data" in value;
    isInstance = isInstance && "links" in value;

    return isInstance;
}

export function ListTransactionsResponseFromJSON(json: any): ListTransactionsResponse {
    return ListTransactionsResponseFromJSONTyped(json, false);
}

export function ListTransactionsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListTransactionsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(TransactionResourceFromJSON)),
        'links': ListAccountsResponseLinksFromJSON(json['links']),
    };
}

export function ListTransactionsResponseToJSON(value?: ListTransactionsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'data': ((value.data as Array<any>).map(TransactionResourceToJSON)),
        'links': ListAccountsResponseLinksToJSON(value.links),
    };
}